requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
colorama==0.4.4
discord.py==1.7.3
et-xmlfile==1.1.0
idna==3.2
iniconfig==1.1.1
multidict==5.1.0
numpy==1.21.2
opencv-python==4.5.3.56
openpyxl==3.0.8
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytesseract==0.3.8
pytest==6.2.5
pytest-asyncio==0.15.1
python-dotenv==0.19.0
requests==2.26.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
